+++ src/Libplanet.Net/Swarm.cs
@@ -999,7 +999,7 @@ namespace Libplanet.Net
             [EnumeratorCancellation] CancellationToken cancellationToken = default
         )
         {
-            BlockLocator locator = blockChain.GetBlockLocator(Options.BranchpointThreshold);
+            BlockLocator locator = blockChain.GetBlockLocator();
             var exceptions = new List<Exception>();
             foreach ((BoundPeer peer, IBlockExcerpt excerpt) in peersWithExcerpts)
             {
@@ -1118,32 +1118,9 @@ namespace Libplanet.Net
                             );
                         }
 
-                        locator = BlockLocator.Create(
-                            startIndex: branchingIndex + downloaded.Count,
-                            idx =>
-                            {
-                                long arg = idx;
-                                if (idx <= branchingIndex)
-                                {
-                                    return blockChain.Store.IndexBlockHash(blockChain.Id, idx);
-                                }
-
-                                int relIdx = (int)(idx - branchingIndex - 1);
-
-                                try
-                                {
-                                    return downloaded[relIdx];
-                                }
-                                catch (ArgumentOutOfRangeException e)
-                                {
-                                    const string msg =
-                                        "Failed to look up a block hash by its index {Index} " +
-                                        "(branching index: {BranchingIndex}; " +
-                                        "downloaded: {Downloaded})";
-                                    _logger.Error(e, msg, arg, branchingIndex, downloaded.Count);
-                                    return null;
-                                }
-                            });
+                        locator = downloaded.Count > 0
+                            ? BlockLocator.Create(tipHash: downloaded.Last())
+                            : locator;
                     }
                     while (downloaded.Count < chunkBlockHashesToDownload);
                 }